fixed deviantt presence effect (friendly npcs take more damage) still existing without the debuff icon mutant eye/sparkling adoration love rays use local iframes (will not interfere with your attacks anymore) wof can't deathray for a second after entering p3 (should never instant deathray in singleplayer anymore) sand elemental bonus life removed and spawns bonus sandnados slower added recipe for heart chocolate flying dutchman nerfed, longer delay between bullet barrages and bullet barrages are much shorter twin rangs will always produce one burst of projectiles upon hitting an enemy or solid tile dubious circuitry probes:  probes and lasers now do 35 damage (was 50 and 80(??????))  probes can fire lasers twice as often  manual probe only fires when you are using a weapon and its not a mining tool  lasers cannot break larvae(?) new mutant's gift description sibling weapons:  have new recipes  sparkling love, styx gazer no longer dropped by devi/abom  devi/abom drop a new material on nohit, no longer first attempt locked  mutant drops material on nohit or 10% chance eye of cthulhu always drops 5 life crystals plantera always drops 5 life fruits deviantt  more delay between teleports on rune wizard attack  random velocity for spectres  light orbs are denser when splitting increased skull charm drop rate to 5% (was 2%) golem  boulders fade in much faster  geysers spawn on ceiling with boulders, other stomp produces floor geysers removed bats spawning in hordes spectre healing lasts longer than before and moves a bit faster again fixed(?) ice queens spawning multiple sets of super flockos ice queen super flockos fire tight spreads of frost waves, but fire slower than before nerfed martian saucer core laser spray duration and random spread prime spin dark star timeleft reduced, but MUCH denser than before twins  retinazer burning aura shrinks while doing deathray  retinazer aura inflicts oiled and on fire too  fixed the despawn meme post plantera dungeon armored bones inherit angry bones bone spray and baby guardians cultist's nebula spheres now rubberband (only speed up when they're a certain distance away from you) active summon weapons:  flesh hand, destroyer gun max requirement reduced to 3  fish stick max reduced to 4  all tooltips now state how many empty slots to get their full power stinger is no longer an ammo added echdeath

+++ Projectiles/Masomode/QueenFlocko.cs
@@ -23,7 +23,7 @@ namespace FargowiltasSouls.Projectiles.Masomode
             }
 
             NPC npc = Main.npc[(int)projectile.ai[0]];
-            if (!(npc.active && npc.type == NPCID.IceQueen && npc.ai[0] != 2))
+            if (!(npc.active && npc.type == NPCID.IceQueen && npc.whoAmI == NPC.FindFirstNPC(NPCID.IceQueen) && npc.ai[0] != 2))
             {
                 projectile.Kill();
                 return;
@@ -49,14 +49,15 @@ namespace FargowiltasSouls.Projectiles.Masomode
                     projectile.velocity *= 1.05f;
             }
 
-            if (++projectile.localAI[0] > 90 && ++projectile.localAI[1] > 60) //fire frost wave
+            if (++projectile.localAI[1] > 90) //fire frost wave
             {
                 projectile.localAI[1] = 0f;
                 Main.PlaySound(SoundID.Item120, projectile.position);
                 if (Main.netMode != NetmodeID.MultiplayerClient)
                 {
                     Vector2 vel = projectile.DirectionTo(player.Center) * 7f;
-                    Projectile.NewProjectile(projectile.Center, vel, ProjectileID.FrostWave, projectile.damage, projectile.knockBack, projectile.owner);
+                    for (int i = -1; i <= 1; i++)
+                        Projectile.NewProjectile(projectile.Center, vel.RotatedBy(MathHelper.ToRadians(5) * i), ProjectileID.FrostWave, projectile.damage, projectile.knockBack, projectile.owner);
                 }
             }
